backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
insightface==0.7.3
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
aiofiles==23.2.1
httpx==0.27.0
reportlab==4.1.0
